Tests/LibWeb/Text/expected/HTML/img-src-change-during-failed-decode.txt

naturalWidth: 1
naturalHeight: 1
